adds probabilistic linear algebra (-l 44 option) and/or multithreading (-t 2) test for make check

+++ test/diff/diff_reals_dim0.sh
@@ -22,4 +22,70 @@ if [ $? -gt 0 ]; then
     exit 4
 fi
 
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -t 2
+if [ $? -gt 0 ]; then
+    exit 21
+fi
+
+diff test/diff/$file.res output_files/$file.res
+if [ $? -gt 0 ]; then
+    exit 22
+fi
+
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -p 256 -t 2
+if [ $? -gt 0 ]; then
+    exit 23
+fi
+
+diff test/diff/$file.res output_files/$file.p256.res
+if [ $? -gt 0 ]; then
+    exit 24
+fi
+
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -l 44
+if [ $? -gt 0 ]; then
+    exit 41
+fi
+
+diff test/diff/$file.res output_files/$file.res
+if [ $? -gt 0 ]; then
+    exit 42
+fi
+
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -p 256 -l 44
+if [ $? -gt 0 ]; then
+    exit 43
+fi
+
+diff test/diff/$file.res output_files/$file.p256.res
+if [ $? -gt 0 ]; then
+    exit 44
+fi
+
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -t 2 -l 44
+if [ $? -gt 0 ]; then
+    exit 61
+fi
+
+diff test/diff/$file.res output_files/$file.res
+if [ $? -gt 0 ]; then
+    exit 62
+fi
+
+$(pwd)/msolve -f input_files/$file.ms -o test/diff/$file.res \
+      -p 256 -t 2 -l 44
+if [ $? -gt 0 ]; then
+    exit 63
+fi
+
+diff test/diff/$file.res output_files/$file.p256.res
+if [ $? -gt 0 ]; then
+    exit 64
+fi
+
 rm test/diff/$file.res
